# Parse HTTP response directly