Was something like this required for #977? (I'm aware tests are failing, will look into it more)

Got the following benchmark results (used the existing benchmark test in `client_test.go`, not sure why delta is nil)

```
name              old time/op    new time/op    delta
Client/4KB-12       62.7µs ± 0%   137.4µs ± 0%   ~     (p=1.000 n=1+1)
Client/50KB-12       119µs ± 0%     132µs ± 0%   ~     (p=1.000 n=1+1)
Client/1000KB-12    1.44ms ± 0%    0.13ms ± 0%   ~     (p=1.000 n=1+1)
Client/2000KB-12    2.68ms ± 0%    0.14ms ± 0%   ~     (p=1.000 n=1+1)

name              old alloc/op   new alloc/op   delta
Client/4KB-12       20.5kB ± 0%    18.2kB ± 0%   ~     (p=1.000 n=1+1)
Client/50KB-12       136kB ± 0%      18kB ± 0%   ~     (p=1.000 n=1+1)
Client/1000KB-12    2.11MB ± 0%    0.02MB ± 0%   ~     (p=1.000 n=1+1)
Client/2000KB-12    4.21MB ± 0%    0.02MB ± 0%   ~     (p=1.000 n=1+1)

name              old allocs/op  new allocs/op  delta
Client/4KB-12         74.0 ± 0%     128.0 ± 0%   ~     (p=1.000 n=1+1)
Client/50KB-12         101 ± 0%       137 ± 0%   ~     (p=1.000 n=1+1)
Client/1000KB-12       595 ± 0%       135 ± 0%   ~     (p=1.000 n=1+1)
Client/2000KB-12     1.10k ± 0%     0.14k ± 0%   ~     (p=1.000 n=1+1)
```

Thank you!