@bboreham 2022-07-26
No, I don't expect the code to get much shorter when implementing this.
In particular code like this which you removed: https://github.com/prometheus/client_golang/pull/1090/files#diff-aa9bfd1a638fbb706f8e8920297902937011160319d9679add5dca56e5ab8277L145-L153
is important to handle timeouts, and will need some functional replacement.

Possibly calling `resp.Body.Close()` will still work, but I think you will need a goroutine watching for context cancellation in order to do this.