@importhuman 2022-08-11
@bboreham will this work? I didn't use a `select` statement since this flow doesn't use a `done` channel. Apologies for the late update.