@importhuman 2022-09-20
Thanks for the detailed feedback! I don't have the answers currently, will update as I look into these issues more.